+++ src/common.ts
@@ -1,4 +1,5 @@
 import { StoreType } from ".";
+import { ObserveTarget, NormalizedObserveTarget, ObservedStoreType } from "./types";
 
 export function noProviderError(): Error {
   return new Error("Wrap your component in a StoreProvider.");
@@ -8,12 +9,14 @@ export function notProvidedError(storeType: StoreType<any>): Error {
   return new Error(`${storeType.name} has not been provided or specified.`);
 }
 
-// @ts-ignore
-export type Instances<T extends StoreType<any>[]> = { [K in keyof T]: InstanceType<T[K]> };
+export function normalizeTarget<T extends ObserveTarget>(target: T): NormalizedObserveTarget<T> {
+  if (typeof target === "function") {
+    return [target as ObservedStoreType<any>, undefined];
+  } else {
+    return target as any;
+  }
+}
 
-export type Omit<T, K extends keyof any> = T extends any ? Pick<T, Exclude<keyof T, K>> : never;
-
-export interface WithStoresProps {
-  useStore: <ST extends StoreType<any>>(storeType: ST) => InstanceType<ST>;
-  useStores: <T extends StoreType<any>[]>(...storeTypes: T) => Instances<T>;
+export function compareArray(arr1: any[], arr2: any[]): boolean {
+  return arr1.length === arr2.length && arr1.every((value, index) => Object.is(value, arr2[index]));
 }