+++ tests/common.tsx
@@ -12,7 +12,7 @@ export class TestStore extends Store<IStore> {
   }
 
   increment() {
-    this.setState({ value: this.state.value + 1 });
+    return this.setState({ value: this.state.value + 1 });
   }
 }
 
@@ -24,5 +24,18 @@ export class AnotherStore extends Store<IAnotherStore> {
   constructor(text: string) {
     super();
     this.state = { text };
+  }
+}
+
+interface IMultiStateStore {
+  state1: string;
+  state2: string;
+  state3: string;
+}
+
+export class MultiStateStore extends Store<IMultiStateStore> {
+  constructor(state1: string, state2: string, state3: string) {
+    super();
+    this.state = { state1, state2, state3 };
   }
 }