+++ tests/useStore.spec.tsx
@@ -1,4 +1,4 @@
-import { TestStore } from "./common";
+import { MultiStateStore, TestStore } from "./common";
 import { mount } from "enzyme";
 import { useStore, StoreProvider } from "../src";
 import React from "react";
@@ -12,12 +12,22 @@ describe("UseStore", () => {
     );
   };
 
+  class UpdateBlocker extends React.PureComponent {
+    render() {
+      return (
+        this.props.children
+      );
+    }
+  }
+
   it("should render with current store state", () => {
     const store = new TestStore(42);
 
     const wrapper = mount(
       <StoreProvider stores={[store]}>
-        <Component />
+        <UpdateBlocker>
+          <Component/>
+        </UpdateBlocker>
       </StoreProvider>,
     );
 
@@ -28,21 +38,23 @@ describe("UseStore", () => {
     const store = new TestStore(42);
 
     // tslint:disable-next-line
-    expect(store["observers"]).toHaveLength(0);
+    expect(store["observers"].size).toBe(0);
 
     const wrapper = mount(
       <StoreProvider stores={[store]}>
+        <UpdateBlocker>
         <Component />
+        </UpdateBlocker>
       </StoreProvider>,
     );
 
     // tslint:disable-next-line
-    expect(store["observers"]).toHaveLength(1);
+    expect(store["observers"].size).toBe(1);
 
     wrapper.unmount();
 
     // tslint:disable-next-line
-    expect(store["observers"]).toHaveLength(0);
+    expect(store["observers"].size).toBe(0);
 
   });
 
@@ -54,12 +66,91 @@ describe("UseStore", () => {
 
     expect(() => mount(
       <StoreProvider stores={[]}>
+        <UpdateBlocker>
         <Component />
+        </UpdateBlocker>
       </StoreProvider>,
     )).toThrowError();
   });
 
   it("should report error with no StoreProvider", () => {
     expect(() => mount(<Component />)).toThrowError();
+  });
+
+  it("should not update when updating a not dependent state", () => {
+
+    const store = new MultiStateStore("state1", "state2", "state3");
+
+    const Component = () => {
+
+      const store = useStore(MultiStateStore, ["state1", (s) => s.state3 ]);
+
+      return (
+        <div>
+          <span id="state1">{store.state.state1}</span>
+          <span id="state2">{store.state.state2}</span>
+          <span id="state3">{store.state.state3}</span>
+        </div>
+      );
+    };
+
+    const wrapper = mount(
+      <StoreProvider stores={[store]}>
+        <UpdateBlocker>
+          <Component />
+        </UpdateBlocker>
+      </StoreProvider>,
+    );
+
+    const expectValues = (state1: string, state2: string, state3: string) => {
+      expect(wrapper.find("#state1").text()).toBe(state1);
+      expect(wrapper.find("#state2").text()).toBe(state2);
+      expect(wrapper.find("#state3").text()).toBe(state3);
+    };
+
+    expectValues("state1", "state2", "state3");
+
+    store.setState({ state2: "123" });
+
+    expectValues("state1", "state2", "state3");
+
+    store.setState({ state3: "new state3"});
+
+    expectValues("state1", "123", "new state3");
+
+  });
+
+  it("should not update when the custom comparer returns true", async () => {
+    const store = new TestStore(42);
+
+    const Component = () => {
+
+      const store = useStore(TestStore, (prev, curr) => prev.value === curr.value - 1);
+
+      return (
+          <span>{store.state.value}</span>
+      );
+    };
+
+    const wrapper = mount(
+      <StoreProvider stores={[store]}>
+        <UpdateBlocker>
+          <Component />
+        </UpdateBlocker>
+      </StoreProvider>,
+    );
+
+    expect(wrapper.find("span").text()).toBe("42");
+
+    await store.increment();
+    wrapper.update();
+
+    expect(wrapper.find("span").text()).toBe("42");
+
+    await store.setState((s) => ({ value: s.value }));
+    wrapper.update();
+
+    expect(wrapper.find("span").text()).toBe("43");
+
   });
 });