+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="2.0.0"></a>
+## [2.0.0](https://github.com/viccrubs/simstate/compare/v1.3.1...v2.0.0) (2019-02-24)
+
+### Features
+
+* Partial Observer feature. Check out `README.md` for usage.
+* Simpler and more consistent APIs for render-props and HOC
+
+### Breaking Changes
+
+* API changes for render-props
+    - Only inject `useStore` into the children
+    - No need to pass in store types; Use the `useStore` directly in children
+* `useStores` are deprecated and will be removed in a future release
+
+
 <a name="1.3.1"></a>
 ## [1.3.1](https://github.com/viccrubs/simstate/compare/v1.3.0...v1.3.1) (2019-02-13)